Reorganization of integral files
