+++ src/Fermi.jl
@@ -27,7 +27,7 @@ include("Backend/PhysicalConstants.jl")
 include("Core/DIIS.jl")
 include("Core/Molecule.jl")                               
 include("Core/Orbitals.jl")
-include("Core/IntegralHelper.jl")
+include("Core/Integrals/IntegralHelper.jl")
 include("Methods/HartreeFock/HartreeFock.jl")
 include("Methods/MollerPlesset/MollerPlesset.jl")
 include("Methods/CoupledCluster/CoupledCluster.jl")
